refactor: update to use adaptive cards in bot-me template (#10714)
* refactor: first draft

* refactor: more

* refactor: ts

* refactor: minor

* refactor: minor
