+++ templates/js/default-bot-message-extension/appPackage/manifest.json.tpl
@@ -88,23 +88,6 @@
                     ]
                 },
                 {
-                    "id": "shareMessage",
-                    "context": [
-                        "message"
-                    ],
-                    "description": "Test command to run action on message context (message sharing)",
-                    "title": "Share Message",
-                    "type": "action",
-                    "parameters": [
-                        {
-                            "name": "includeImage",
-                            "title": "Include Image",
-                            "description": "Include image in Hero Card",
-                            "inputType": "toggle"
-                        }
-                    ]
-                },
-                {
                     "id": "searchQuery",
                     "context": [
                         "compose",
