+++ templates/js/default-bot-message-extension/package.json.tpl
@@ -10,16 +10,18 @@
     },
     "author": "Microsoft",
     "license": "MIT",
-    "main": "index.js",
+    "main": "src/index.js",
     "scripts": {
         "dev:teamsfx": "env-cmd --silent -f .localConfigs npm run dev",
-        "dev": "nodemon --inspect=9239 --signal SIGINT ./index.js",
-        "start": "node ./index.js",
-        "watch": "nodemon ./index.js",
+        "dev": "nodemon --inspect=9239 --signal SIGINT ./src/index.js",
+        "start": "node ./src/index.js",
+        "watch": "nodemon ./src/index.js",
         "test": "echo \"Error: no test specified\" && exit 1"
     },
-    "dependencies": {
-        "@microsoft/adaptivecards-tools": "^1.0.0",
+   "dependencies": {
+        "adaptive-expressions": "^4.20.0",
+        "adaptivecards-templating": "^2.3.1",
+        "adaptivecards": "^3.0.1",
         "botbuilder": "^4.20.0",
         "restify": "^10.0.0"
     },
