+++ templates/js/default-bot-message-extension/src/adaptiveCards/learn.json
