+++ templates/js/default-bot-message-extension/src/adaptiveCards/welcome.json
