+++ templates/js/default-bot-message-extension/src/config.js
