+++ templates/js/default-bot-message-extension/src/index.js
