+++ templates/ts/default-bot-message-extension/src/adaptiveCards/actionCard.json
@@ -1,0 +1,26 @@
+
+{
+    "type": "AdaptiveCard",
+    "body": [
+        {
+            "type": "TextBlock",
+            "text": "${title}",
+            "wrap": true,
+            "size": "Large"
+          },
+          {
+            "type": "TextBlock",
+            "text": "${subTitle}",
+            "wrap": true,
+            "size": "Medium"
+          },
+          {
+            "type": "TextBlock",
+            "text": "${text}",
+            "wrap": true,
+            "size": "Small"
+          }
+    ],
+    "$schema": "http://adaptivecards.io/schemas/adaptive-card.json",
+    "version": "1.4"
+}
