+++ templates/ts/default-bot-message-extension/src/adaptiveCards/learn.json
