+++ templates/ts/default-bot-message-extension/src/adaptiveCards/welcome.json
