+++ templates/ts/default-bot-message-extension/web.config
@@ -11,13 +11,13 @@
     <webSocket enabled="false" />
     <handlers>
       <!-- Indicates that the server.js file is a node.js site to be handled by the iisnode module -->
-      <add name="iisnode" path="lib/index.js" verb="*" modules="iisnode"/>
+      <add name="iisnode" path="lib/src/index.js" verb="*" modules="iisnode"/>
     </handlers>
     <rewrite>
       <rules>
         <!-- Do not interfere with requests for node-inspector debugging -->
         <rule name="NodeInspector" patternSyntax="ECMAScript" stopProcessing="true">
-          <match url="^lib/index.js\/debug[\/]?" />
+          <match url="^lib/src/index.js\/debug[\/]?" />
         </rule>
 
         <!-- First we consider whether the incoming URL matches a physical file in the /public folder -->
@@ -30,7 +30,7 @@
           <conditions>
             <add input="{REQUEST_FILENAME}" matchType="IsFile" negate="True"/>
           </conditions>
-          <action type="Rewrite" url="lib/index.js"/>
+          <action type="Rewrite" url="lib/src/index.js"/>
         </rule>
       </rules>
     </rewrite>
